requirements.txt
flask==3.0.0
twilio==8.10.0
python-dotenv==1.0.0
pandas==2.1.3
requests==2.31.0
openai>=1.40.0
httpx>=0.25.0
python-decouple==3.8
gunicorn==21.2.0

